requirements.txt
cfn-flip==1.2.2
click==7.1.1
PyYAML==5.3.1
six==1.14.0
troposphere==2.6.0
